+++ packages/convostack-agent-sse/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @convostack/agent-sse
 
+## 0.0.11
+
+### Patch Changes
+
+- Misc frontend updates, adding postgres and mysql to convostack
+- Updated dependencies
+  - @convostack/agent@0.0.11
+  - @convostack/shared@0.0.11
+
 ## 0.0.10
 
 ### Patch Changes